+++ lib/assets.php
@@ -3,88 +3,78 @@
  * Scripts and stylesheets
  *
  * Enqueue stylesheets in the following order:
- * 1. /theme/assets/dist/css/main.css
+ * 1. /theme/dist/styles/main.css
  *
  * Enqueue scripts in the following order:
- * 1. jquery-2.1.1.min.js via Google CDN
- * 2. /theme/assets/dist/js/modernizr.min.js
- * 3. /theme/assets/dist/js/scripts.js
+ * 1. jquery.js via Google CDN
+ * 2. /theme/dist/scripts/modernizr.js
+ * 3. /theme/dist/scripts/app.js
  *
  * Google Analytics is loaded after enqueued scripts if:
  * - An ID has been defined in config.php
  * - You're not logged in as an administrator
  */
-function forward_asset_path($filename_dev, $filename) {
+function roots_asset_path($filename) {
   if (WP_ENV === 'development') {
-    return get_template_directory_uri() . '/assets/dist/' . $filename_dev;
+    return get_template_directory_uri() . '/dist/' . $filename;
   }
 
-  $manifest_path = get_template_directory() . '/assets/dist/rev-manifest.json';
-
+  $manifest_path = get_template_directory() . '/dist/rev-manifest.json';
   if (file_exists($manifest_path)) {
     $manifest = json_decode(file_get_contents($manifest_path), true);
   } else {
     $manifest = [];
   }
-
   if (array_key_exists($filename, $manifest)) {
-    return get_template_directory_uri() . '/assets/dist/' . $manifest[$filename];
+    return get_template_directory_uri() . '/dist/' . $manifest[$filename];
   }
 }
 
-function forward_assets() {
-  wp_enqueue_style('forward_css', forward_asset_path('css/main.css', 'css/main.min.css'), false, null);
 
+function roots_assets() {
+  wp_enqueue_style('roots_css', roots_asset_path('styles/main.css'), false, null);
   /**
-   * jQuery is loaded using the same method from HTML5 Boilerplate:
    * Grab Google CDN's latest jQuery with a protocol relative URL; fallback to local if offline
-   * It's kept in the header instead of footer to avoid conflicts with plugins.
+   * jQuery & Modernizr load in the footer per HTML5 Boilerplate's recommendation: http://goo.gl/nMGR7P
+   * If a plugin enqueues jQuery-dependent scripts in the head, jQuery will load in the head to meet the plugin's dependencies
+   * To explicitly load jQuery in the head, change the last wp_enqueue_script parameter to false
    */
   if (!is_admin() && current_theme_supports('jquery-cdn')) {
     wp_deregister_script('jquery');
-
-    if (WP_ENV === 'development') {
-      wp_register_script('jquery', '//ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.js', array(), null, true);
-    } else {
-      wp_register_script('jquery', '//ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js', array(), null, true);
-    }
-
-    add_filter('script_loader_src', 'forward_jquery_local_fallback', 10, 2);
+    wp_register_script('jquery', '//ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.js', array(), null, true);
+    add_filter('script_loader_src', 'roots_jquery_local_fallback', 10, 2);
   }
-
   if (is_single() && comments_open() && get_option('thread_comments')) {
     wp_enqueue_script('comment-reply');
   }
-
-  wp_enqueue_script('modernizr', forward_asset_path('../vendor/modernizr/modernizr.js', 'js/modernizr.min.js'), array(), null, true);
+  wp_enqueue_script(roots_asset_path('scripts/modernizr.js'), array(), null, true);
   wp_enqueue_script('jquery');
-  wp_enqueue_script('forward_js', forward_asset_path('js/scripts.js', 'js/scripts.min.js'), array(), null, true);
+  wp_enqueue_script('roots_js', roots_asset_path('scripts/app.js'), array(), null, true);
 }
-add_action('wp_enqueue_scripts', 'forward_assets', 100);
+add_action('wp_enqueue_scripts', 'roots_assets', 100);
+
 
 // http://wordpress.stackexchange.com/a/12450
-function forward_jquery_local_fallback($src, $handle = null) {
+function roots_jquery_local_fallback($src, $handle = null) {
   static $add_jquery_fallback = false;
-
   if ($add_jquery_fallback) {
-    echo '<script>window.jQuery || document.write(\'<script src="' . get_template_directory_uri() . '/assets/dist/js/jquery-2.1.1.min.js"><\/script>\')</script>' . "\n";
+    echo '<script>window.jQuery || document.write(\'<script src="' . get_template_directory_uri() . '/dist/scripts/jquery.js"><\/script>\')</script>' . "\n";
     $add_jquery_fallback = false;
   }
-
   if ($handle === 'jquery') {
     $add_jquery_fallback = true;
   }
-
   return $src;
 }
-add_action('wp_head', 'forward_jquery_local_fallback');
+add_action('wp_head', 'roots_jquery_local_fallback');
+
 
 /**
  * Google Analytics snippet from HTML5 Boilerplate
  *
  * Cookie domain is 'auto' configured. See: http://goo.gl/VUCHKM
  */
-function forward_google_analytics() { ?>
+function roots_google_analytics() { ?>
 <script>
   <?php if (WP_ENV === 'production') : ?>
     (function(b,o,i,l,e,r){b.GoogleAnalyticsObject=l;b[l]||(b[l]=
@@ -102,5 +92,5 @@ function forward_google_analytics() { ?>
 
 <?php }
 if (GOOGLE_ANALYTICS_ID && (WP_ENV !== 'production' || !current_user_can('manage_options'))) {
-  add_action('wp_footer', 'forward_google_analytics', 20);
+  add_action('wp_footer', 'roots_google_analytics', 20);
 }